backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
playwright
beautifulsoup4
lxml
celery
redis
phonenumbers
email-validator
geopy
psycopg2-binary
python-multipart
alembic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
httpx
dnspython
requests
